# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BookWorm API
  description: |
    Full-stack online bookstore REST API.
    
    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    
    ```
    Authorization: Bearer <token>
    ```
    
    ## Getting Started
    1. Register at `POST /auth/register` or login at `POST /auth/login`
    2. Copy the returned `token`
    3. Click "Authorize" above and paste: `Bearer <token>`
  version: 1.0.0
  contact:
    name: BookWorm API Support
servers:
  - url: http://localhost:5000/api/v1
    description: Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        is_guest:
          type: boolean
        gift_points:
          type: integer
        created_at:
          type: string
          format: date-time
    Book:
      type: object
      properties:
        book_id:
          type: string
          format: uuid
        title:
          type: string
        author_name:
          type: string
        category_name:
          type: string
        price:
          type: number
        format:
          type: string
          enum: [Paperback, Hardcover, eBook]
        isbn:
          type: string
        description:
          type: string
        cover_image_url:
          type: string
        stock_quantity:
          type: integer
        is_bestseller:
          type: boolean
        is_new_launch:
          type: boolean
        is_recommended:
          type: boolean
        avg_rating:
          type: number
        review_count:
          type: integer
...
    Address:
      type: object
      properties:
        address_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        address_line:
          type: string
        city:
          type: string
        state:
          type: string
        pin_code:
          type: string
        country:
          type: string
        phone_number:
          type: string
        is_default:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, first_name, last_name]
              properties:
                email:
                  type: string
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                first_name:
                  type: string
                last_name:
                  type: string
                phone_number:
                  type: string
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: alice@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
...
  /auth/verify:
    get:
      tags: [Authentication]
      summary: Verify JWT token validity
      responses:
        '200':
          description: Token is valid
        '401':
          description: Invalid token

  # ── Users ───────────────────────────────────────────────────────────
...
  /books:
    get:
      tags: [Books]
      summary: Get all books with optional filters
      security: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
            format: uuid
        - in: query
          name: bestseller
          schema:
            type: boolean
        - in: query
          name: new_launch
          schema:
            type: boolean
        - in: query
          name: recommended
          schema:
            type: boolean
        - in: query
          name: sort
          schema:
            type: string
            enum: [price_asc, price_desc, bestseller]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: List of books with pagination
  /books/search:
    get:
      tags: [Books]
      summary: Search books by title or author
      security: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
  /books/{bookId}:
    get:
      tags: [Books]
      summary: Get book details
      security: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
  /books/{bookId}/related:
    get:
      tags: [Books]
      summary: Get related books
      security: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Related books
  /books/{bookId}/reviews:
    get:
      tags: [Books]
      summary: Get book reviews
      security: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book reviews

  # ── Categories ──────────────────────────────────────────────────────
...
  /categories/{categoryId}/books:
    get:
      tags: [Categories]
      summary: Get books by category
      security: []
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Books in category

  # ── Authors ─────────────────────────────────────────────────────────
  /authors/{authorId}:
    get:
      tags: [Authors]
      summary: Get author details
      security: []
      parameters:
        - in: path
          name: authorId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Author details
  /authors/{authorId}/books:
    get:
      tags: [Authors]
      summary: Get books by author
      security: []
      parameters:
        - in: path
          name: authorId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Author's books

  # ── Cart ────────────────────────────────────────────────────────────
...
  /addresses:
    get:
      tags: [Addresses]
      summary: Get all user addresses
      responses:
        '200':
          description: List of addresses
    post:
      tags: [Addresses]
      summary: Create new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        '201':
          description: Address created
  /addresses/{addressId}:
    put:
      tags: [Addresses]
      summary: Update address
      parameters:
        - in: path
          name: addressId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Address updated
    delete:
      tags: [Addresses]
      summary: Delete address
      parameters:
        - in: path
          name: addressId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Address deleted
  /addresses/{addressId}/set-default:
    put:
      tags: [Addresses]
      summary: Set address as default
      parameters:
        - in: path
          name: addressId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Default address updated

  # ── Orders ──────────────────────────────────────────────────────────
...